add prepare scripts

+++ 00-prepare-vm.sh
@@ -8,6 +8,12 @@ if [ $? -ne 0 ]; then
     clear && echo 'problems with changing users' && exit 1
 fi
 
+sed 's~metalink=https://mirrors.centos.org/metalink?repo=centos-baseos-$stream&arch=$basearch&protocol=https,http~baseurl=https://mirror.yandex.ru/centos-stream/9-stream/BaseOS/x86_64/os/~g' -i /etc/yum.repos.d/centos.repo && sed 's~metalink=https://mirrors.centos.org/metalink?repo=centos-appstream-$stream&arch=$basearch&protocol=https,http~baseurl=https://mirror.yandex.ru/centos-stream/9-stream/AppStream/x86_64/os/~g' -i /etc/yum.repos.d/centos.repo && sed 's~metalink=https://mirrors.centos.org/metalink?repo=centos-crb-$stream&arch=$basearch&protocol=https,http~baseurl=https://mirror.yandex.ru/centos-stream/9-stream/CRB/x86_64/os/~g' -i /etc/yum.repos.d/centos.repo
+
+if [ $? -ne 0 ]; then
+    clear && echo 'problems with patching repository lists' && exit 1
+fi
+
 wget https://dl.fedoraproject.org/pub/epel/epel-release-latest-9.noarch.rpm && yum install -y epel-release-latest-9.noarch.rpm && rm -f epel-release-latest-9.noarch.rpm
 
 if [ $? -ne 0 ]; then
@@ -18,12 +24,6 @@ wget https://dl.fedoraproject.org/pub/epel/epel-next-release-latest-9.noarch.rpm
 
 if [ $? -ne 0 ]; then
     clear && echo 'problems adding epel 9 repository' && exit 1
-fi
-
-sed 's~metalink=https://mirrors.centos.org/metalink?repo=centos-baseos-$stream&arch=$basearch&protocol=https,http~baseurl=https://mirror.yandex.ru/centos-stream/9-stream/BaseOS/x86_64/os/~g' -i /etc/yum.repos.d/centos.repo && sed 's~metalink=https://mirrors.centos.org/metalink?repo=centos-appstream-$stream&arch=$basearch&protocol=https,http~baseurl=https://mirror.yandex.ru/centos-stream/9-stream/AppStream/x86_64/os/~g' -i /etc/yum.repos.d/centos.repo && sed 's~metalink=https://mirrors.centos.org/metalink?repo=centos-crb-$stream&arch=$basearch&protocol=https,http~baseurl=https://mirror.yandex.ru/centos-stream/9-stream/CRB/x86_64/os/~g' -i /etc/yum.repos.d/centos.repo
-
-if [ $? -ne 0 ]; then
-    clear && echo 'problems with patching repository lists' && exit 1
 fi
 
 yum install -y iproute iputils net-tools ncurses sshpass pv nano expect dash bc
@@ -53,4 +53,4 @@ cat /usr/bin/deploy.vt- | base64 -d > /usr/bin/deploy.vt && rm -f /usr/bin/deplo
 
 chmod +x /labs/**/*.sh
 
-echo "Done, now log in as 'moon' with password 'selena'.
+echo "Done, now log in as 'moon' with password 'selena'."
